+++ src/xz/lasercutter/PropertyManager.java
@@ -1,126 +1,173 @@
 package xz.lasercutter;
 
-import javax.swing.JFrame;
+import java.io.*;
+import java.util.Properties;
 
-public class PropertyManager extends JFrame {
-	public static final int MOTOR_MOVE_DISTANCE = 1400;
-	public static final int MOTOR_MOVE_MARGIN = 25;
-	public static final int MOTOR_TURNING_EPS_Y = 15; //vertical
-	public static final int MOTOR_TURNING_EPS_X = 9; //horizontal
+public class PropertyManager {
+	public static final int MOTOR_MOVE_DISTANCE 	= 1400;
+	public static final int MOTOR_MOVE_MARGIN 		= 25;
+	public static final int MOTOR_TURNING_EPS_Y 	= 15; //vertical
+	public static final int MOTOR_TURNING_EPS_X 	= 9; //horizontal
 	
-	public static final String[] POSSIBLE_PORT_NAMES = {
-		"/dev/ttyUSB0",
-		"COM5",
-		"COM4",
-		"/dev/ttyUSB1",
-		"/dev/ttyUSB2"
-	};
+	// Constants 
+	private static final String directoryTempFiles 	= "./tmp/";
+	private static final String fileNameConvPic 	= "converted_pic.png";
+	private static final String fileNameConfig 		= "config.properties";
+	private static final String fileNameLog 		= "log.txt";
+	private static final String fileNameCmdLst 		= "laser_cmd.txt";
 	
-	private static String portName;
-	//Directory
-	private static String initialPicDirectory = "./";
-	private static String tempDirectory = "./tmp/";
-	private static String tempPicFileName = "converted_pic.png";
-	private static String tempLogFileName = "log.txt";
-	private static String tempCmdFileName = "laser_cmd.txt";
+	// Default value of variables
+	private static String defaultPortName 			= "";
+	private static String defaultPossiblePortNames 	= "";
+	private static String defaultInitPicDir 		= ".";
+	private static int defaultDrawBrightness 		= 3;
+	private static int defaultLineDelay 			= 50;
+	private static int defaultDotDelay 				= 500;
+	private static int defaultChoicedPrintMethod 	= ImageConverter.NUMBER_OF_PRINT_METHODS - 1;
 	
-	private static int drawBrightness = 3;
-	private static int drawLineDelay = 50; // draft paper:50, do not set this value less than 50
-	private static int drawDotDelay = 500;  // draft paper:90+
+	// Property
+	private static Properties properties = new Properties();
 	
-	public static String getTempLogFileName() {
-		return tempLogFileName;
+	static {
+		loadProperties();
 	}
-
-	public static void setTempLogFileName(String tempLogFileName) {
-		PropertyManager.tempLogFileName = tempLogFileName;
+	
+	public static void loadProperties() {
+		String filePathConfig = directoryTempFiles + fileNameConfig;
+		File fileConfig = null;
+		FileInputStream configInputStream = null;
+		
+		properties.clear();
+		
+		try {
+			fileConfig = new File(filePathConfig);
+			configInputStream = new FileInputStream(fileConfig);
+			properties.load(configInputStream);
+		} catch (IOException e) {
+			MainWindow.log("SYSTEM\t|Configure file not found.");
+		} finally {
+			if (configInputStream != null)
+				try {
+					configInputStream.close();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+		}
+		
 	}
-
+	
+	public static void storeProperties() {
+		String filePathConfig = directoryTempFiles + fileNameConfig;
+		File fileConfig = null;
+		FileOutputStream configOutputStream = null;
+		
+		try {
+			fileConfig = new File(filePathConfig);
+			fileConfig.createNewFile();
+			configOutputStream = new FileOutputStream(fileConfig);
+			properties.store(configOutputStream, null);
+		} catch (IOException e) {
+			MainWindow.log("SYSTEM\t|Configure file writing error.");
+		} finally {
+			if (configOutputStream != null)
+				try {
+					configOutputStream.close();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+		}
+		
+	}
+	
+	public static String getLogFilePath() {
+		return directoryTempFiles + fileNameLog;
+	}
+	public static String getConvPicFilePath() {
+		return directoryTempFiles + fileNameConvPic;
+	}
+	public static String getCmdLstFilePath() {
+		return directoryTempFiles + fileNameCmdLst;
+	}
+	
 	public static int getDrawBrightness() {
-		return drawBrightness;
-	}
-
-	public static void setDrawBrightness(int drawBrightness) {
-		PropertyManager.drawBrightness = drawBrightness;
+		String s = properties.getProperty("DrawBrightness");
+		if (s == null) {
+			properties.setProperty("DrawBrightness", "" + defaultDrawBrightness);
+			s = "" + defaultDrawBrightness;
+		}
+		return Integer.parseInt(s);
 	}
 
 	public static int getDrawLineDelay() {
-		return drawLineDelay;
-	}
-
-	public static void setDrawLineDelay(int drawLineDelay) {
-		PropertyManager.drawLineDelay = drawLineDelay;
+		String s = properties.getProperty("LineDelay");
+		if (s == null) {
+			properties.setProperty("LineDelay", "" + defaultLineDelay);
+			s = "" + defaultLineDelay;
+		}
+		return Integer.parseInt(s);
 	}
 
 	public static int getDrawDotDelay() {
-		return drawDotDelay;
-	}
-
-	public static void setDrawDotDelay(int drawDotDelay) {
-		PropertyManager.drawDotDelay = drawDotDelay;
-	}
-
-	public static String getTempLogPath() {
-		return tempDirectory + tempLogFileName;
-	}
-	public static String getTempPicPath() {
-		return tempDirectory + tempPicFileName;
+		String s = properties.getProperty("DotDelay");
+		if (s == null) {
+			properties.setProperty("DotDelay", "" + defaultDotDelay);
+			s = "" + defaultDotDelay;
+		}
+		return Integer.parseInt(s);
 	}
 	
-	public static String getTempCmdPath() {
-		return tempDirectory + tempCmdFileName;
+	public static int getChoicedPrintMethod() {
+		String s = properties.getProperty("ChoicedPrintMethod");
+		if (s == null) {
+			properties.setProperty("ChoicedPrintMethod", "" + defaultChoicedPrintMethod);
+			s = "" + defaultChoicedPrintMethod;
+		}
+		int x = Integer.parseInt(s);
+		if (x < 0 || x >= ImageConverter.NUMBER_OF_PRINT_METHODS)
+			x = defaultChoicedPrintMethod;
+		return x;
 	}
 	
-	public static String getTempPicFileName() {
-		return tempPicFileName;
+	public static void setChoicedPrintMethod(int cpm) {
+		properties.setProperty("ChoicedPrintMethod", "" + cpm);
+		storeProperties();
 	}
-
-	public static void setTempPicFileName(String tempPicFileName) {
-		PropertyManager.tempPicFileName = tempPicFileName;
+	
+	public static String getInitialImageDirectory() {
+		String s = properties.getProperty("InitialPictureDirectory");
+		if (s == null) {
+			properties.setProperty("InitialPictureDirectory", defaultInitPicDir);
+			s = defaultInitPicDir;
+		}
+		return s;
 	}
-
-
-	public static String getTempCmdFileName() {
-		return tempCmdFileName;
+	
+	public static void setInitialImageDirectory(String iid) {
+		properties.setProperty("InitialPictureDirectory", iid);
+		storeProperties();
 	}
-
-
-	public static void setTempCmdFileName(String tempCmdFileName) {
-		PropertyManager.tempCmdFileName = tempCmdFileName;
+	
+	public static String[] getPossiblePortNames() {
+		String s = properties.getProperty("PossiblePortNames");
+		if (s == null) {
+			properties.setProperty("PossiblePortNames", defaultPossiblePortNames);
+			s = defaultPossiblePortNames;
+		}
+		return (getPortName() + "," + s).split(",");
 	}
-
-
-	public static String getTempPath() {
-		return tempDirectory;
-	}
-
-
-	public static void setTempPath(String tempPath) {
-		PropertyManager.tempDirectory = tempPath;
-	}
-
-
-	public static String getInitialPicPath() {
-		return initialPicDirectory;
-	}
-
-
-	public static void setInitialPicPath(String initialPicPath) {
-		PropertyManager.initialPicDirectory = initialPicPath;
-	}
-
-
+	
 	public static String getPortName() {
-		return portName;
+		String s = properties.getProperty("PortName");
+		if (s == null) {
+			properties.setProperty("PortName", defaultPortName);
+			s = defaultPortName;
+		}
+		return s;
 	}
-
-	public static void setPortName(String portName) {
-		PropertyManager.portName = portName;
-		MainWindow.setPortName(portName);
+	
+	public static void setPortName(String pn) {
+		properties.setProperty("PortName", pn);
+		MainWindow.updatePortName(pn);
+		storeProperties();
 	}
-
-
-	public PropertyManager() {
-	}
-
 }